fix(query): correct geo_distance, nested envelope; add inner_hits
GeoDistance was emitting {pin: {location: ...}} which Elasticsearch
rejects — fixed to emit the proper geo_distance envelope with the
required distance argument, plus distance_type/validation_method/
ignore_unmapped/boost. The old test asserted the broken shape, so
the bug shipped silently.

Nested was wrapping query in an extra array level (query: [bool: ...]
instead of query: bool: ...) which Elasticsearch also rejects.
Added score_mode/ignore_unmapped/inner_hits. The empty-collection
case now emits an stdClass to keep ES happy.

InnerHits added as a typed sub-object (name, from, size, sort,
_source, highlight, explain, script_fields, docvalue_fields, etc.)
so nested/has_child/has_parent can use it.

PhrasePrefix boost type changed from int to float for consistency
with every other query. GeoDistanceSort ignore_unmapped is now a
constructor arg instead of hard-coded true.

Tests now use AbstractElasticTestCase, indexing real docs through
a typed mapping and asserting the round-tripped hit count.

+++ src/Query/PhrasePrefix.php
@@ -10,7 +10,7 @@ class PhrasePrefix implements \Spameri\ElasticQuery\Query\LeafQueryInterface
 	public function __construct(
 		private string $field,
 		private string $queryString,
-		private int $boost = 1,
+		private float $boost = 1.0,
 		private int $slop = 1,
 	)
 	{
